+++ OKX.Api/Account/Responses/OkxAccountPrecheckSetDeltaNeutral.cs
@@ -1,0 +1,13 @@
+namespace OKX.Api.Account;
+
+/// <summary>
+/// Precheck result for switching delta neutral strategy
+/// </summary>
+public record OkxAccountPrecheckSetDeltaNeutral
+{
+    /// <summary>
+    /// Unmatched information list
+    /// </summary>
+    [JsonProperty("unmatchedInfoCheck")]
+    public List<OkxAccountPrecheckModeUnmatchedInformation>? UnmatchedInformationList { get; set; } = [];
+}